backend/fastapi_backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
celery
redis
python-dotenv